Adding comments

+++ 2-code-challenge/src/main/java/com/forbes/takehometest/controller/DictionaryController.java
@@ -4,6 +4,7 @@ import com.forbes.takehometest.interfaces.IDictionaryService;
 import com.forbes.takehometest.model.dictionary.DictionaryAddModel;
 import com.forbes.takehometest.model.dictionary.DictionaryListModel;
 import com.forbes.takehometest.model.dictionary.DictionaryRemoveModel;
+import com.forbes.takehometest.util.WordValidationUtils;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -42,12 +43,11 @@ public class DictionaryController {
 		}
 		var duplicateWords = new ArrayList<String>();
 		for (var word : addModel.getWordsToAdd()) {
+			word = WordValidationUtils.sanitizeWord(word).toLowerCase();
 			// ignore word if it has digits
-			if (word.matches(".*\\d+.*")) {
+			if (!WordValidationUtils.isValidDictionaryWord(word)) {
 				continue;
 			}
-			// remove leading/trailing spaces and punctuation from the word if present. Convert to lowercase.
-			word = word.trim().replaceAll("^\\p{Punct}", "").replaceAll("\\p{Punct}$", "").toLowerCase();
 			if (!dictionaryService.addWord(word)) {
 				duplicateWords.add(word);
 			}
@@ -59,10 +59,19 @@ public class DictionaryController {
 		return ResponseEntity.accepted().build();
 	}
 
+	/**
+	 * Deletes the given words from the dictionary if valid and present. (Valid means that there's no miscellaneous
+	 * punctuation or digits aside from ' and -.
+	 *
+	 * We do not convert to lowercase here to be safe in case the user was expecting to delete an uppercase word.
+	 */
 	@DeleteMapping("/dictionary")
 	public ResponseEntity<Object> deleteWord(@RequestBody DictionaryRemoveModel removeModel) {
 		var notFoundWords = new ArrayList<String>();
 		for (var word : removeModel.getWordsToRemove()) {
+			if (!WordValidationUtils.isValidDictionaryWord(word)) {
+				continue;
+			}
 			if (!dictionaryService.removeWord(word)) {
 				notFoundWords.add(word);
 			}
